Allow Scanner to work outside a Bundler environment
When a LockfileParser is initialised, Bundler will try to find the
related Gemfile in the current directory’s hierarchy. If the tool is
run from a directory that doesn't have a Gemfile in its hierarchy,
Bundler will raise an exception and make the tool fail.

As we don’t need to do anything with the Gemfile in this tool, we tell
Bundler the location of a (dummy) Gemfile explicitly so it doesn’t go
looking.